Add Dockerfile for production deployment and update .gitignore
- Introduced `Dockerfile.prod` for building a production-ready Docker image, including dependency installation and application setup.
- Updated `.gitignore` to exclude `internal-docs` from version control.
- Modified `package.json` to add new scripts for production and cache clearing.
- Updated `package-lock.json` to reflect changes in dependencies, including the addition of TypeScript.
- Removed obsolete files and adjusted TypeScript configuration for improved build management.

+++ app/views/hlsEventAggregated.ts
@@ -7,14 +7,14 @@ interface HLSEventAggregated {
   appName: string;
   eventType: string;
   stage: string;
-  totalEventsState: number & typia.tags.Type<"int64">;
-  uniqueUsersState: number & typia.tags.Type<"int64">;
-  avgBitrateState: number & typia.tags.Type<"double">;
-  maxBitrateState: number & typia.tags.Type<"int64">;
-  minBitrateState: number & typia.tags.Type<"int64">;
-  levelSwitchesState: number & typia.tags.Type<"int64">;
-  playbackStartsState: number & typia.tags.Type<"int64">;
-  avgFragmentDurationState: number & typia.tags.Type<"double">;
+  totalEventsState: number & typia.tags.Type<"uint64">;
+  uniqueUsersState: number & typia.tags.Type<"uint64">;
+  avgBitrateState: number & typia.tags.Type<"float">;
+  maxBitrateState: number & typia.tags.Type<"uint64">;
+  minBitrateState: number & typia.tags.Type<"uint64">;
+  levelSwitchesState: number & typia.tags.Type<"uint64">;
+  playbackStartsState: number & typia.tags.Type<"uint64">;
+  avgFragmentDurationState: number & typia.tags.Type<"float">;
 }
 
 const hlsTable = HLSEventPipeline.table!;
